複数の API の連続実行 API の失敗
の処理を実装

+++ LibrarySample/VolleySample/app/src/main/java/m/k/android/librarysample/volleysample/model/WindJson.java
@@ -7,6 +7,6 @@ import lombok.Value;
  */
 @Value
 public class WindJson {
-    private int deg;
+    private double deg;
     private double speed;
 }
